fix: address structured output review feedback
